requirements.txt
# NeXOptimIA - Requerimientos
# Producto estrella de OPNeXOX Hardware y Software
# Integra OPNeXOX Chips y OPNeXOX Labs
# Edge AI para Optimización de Redes Eléctricas y más

tensorflow==2.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
xgboost>=1.7.0
lightgbm>=4.0.0

# ============ TIME SERIES ============
tsfresh>=0.20.0
statsmodels>=0.14.0
prophet>=1.1.0

# ============ OPTIMIZATION ============
pulp>=2.7.0
pyomo>=6.6.0
scipy>=1.11.0

# ============ IOT & SENSORS ============
pyserial>=3.5
pymodbus>=3.5.0
paho-mqtt>=1.6.0
adafruit-circuitpython-rfm9x>=2.2.0

# ============ EDGE DEPLOYMENT ============
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
docker>=6.1.0

# ============ DATA PROCESSING ============
dask>=2023.8.0
joblib>=1.3.0
imbalanced-learn>=0.11.0

# ============ VISUALIZATION ============
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ============ CRYPTO & SECURITY ============
pycryptodome>=3.18.0
cryptography>=41.0.0

# ============ TESTING & DEVELOPMENT ============
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0


# ============ ENVIRONMENTAL ============
python-dotenv>=1.0.0


# ============ LINUX/WSL ONLY (Comenta si usas Windows) ============
# tflite-runtime==2.15.0
# pylorawan>=0.2.0
# lora-gateway>=1.0.0
# chirpstack-api>=4.0.0
